backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.2
openai==2.26.0
chromadb==1.5.6
sentence-transformers==5.3.0
pypdf==6.8.0
